@@ -1,8 +1,11 @@
 package cmd
 
 import (
-	"github.com/ErdemOzgen/blackdagger/app"
+	"log"
+
 	"github.com/ErdemOzgen/blackdagger/internal/config"
+	"github.com/ErdemOzgen/blackdagger/internal/frontend"
+	"github.com/ErdemOzgen/blackdagger/internal/logger"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
@@ -12,27 +15,43 @@ func serverCmd() *cobra.Command {
 		Use:   "server",
 		Short: "Start the server",
 		Long:  `blackdagger server [--dags=<DAGs dir>] [--host=<host>] [--port=<port>]`,
-		PreRun: func(cmd *cobra.Command, args []string) {
+		PreRun: func(cmd *cobra.Command, _ []string) {
 			_ = viper.BindPFlag("port", cmd.Flags().Lookup("port"))
 			_ = viper.BindPFlag("host", cmd.Flags().Lookup("host"))
 			_ = viper.BindPFlag("dags", cmd.Flags().Lookup("dags"))
-			cobra.CheckErr(config.LoadConfig(homeDir))
 		},
-		Run: func(cmd *cobra.Command, args []string) {
-			// TODO: move to config files
+		Run: func(cmd *cobra.Command, _ []string) {
+			cfg, err := config.Load()
+			if err != nil {
+				log.Fatalf("Configuration load failed: %v", err)
+			}
+			logger := logger.NewLogger(logger.NewLoggerArgs{
+				Debug:  cfg.Debug,
+				Format: cfg.LogFormat,
+			})
+
+			logger.Info("Server initialization", "host", cfg.Host, "port", cfg.Port)
+
 			pullDagList := []string{"default"}
 			Pulldags(pullDagList)
-			service := app.NewFrontendService()
-			err := service.Start(cmd.Context())
-			checkError(err)
+
+			dataStore := newDataStores(cfg)
+			cli := newClient(cfg, dataStore, logger)
+			server := frontend.New(cfg, logger, cli)
+			if err := server.Serve(cmd.Context()); err != nil {
+				logger.Fatal("Server initialization failed", "error", err)
+			}
 		},
 	}
+
 	bindServerCommandFlags(cmd)
 	return cmd
 }
 
 func bindServerCommandFlags(cmd *cobra.Command) {
-	cmd.Flags().StringP("dags", "d", "", "location of DAG files (default is $HOME/.blackdagger/dags)")
+	cmd.Flags().StringP(
+		"dags", "d", "", "location of DAG files (default is $HOME/.config/blackdagger/dags)",
+	)
 	cmd.Flags().StringP("host", "s", "", "server host (default is localhost)")
 	cmd.Flags().StringP("port", "p", "", "server port (default is 8080)")
 }
