@@ -2,29 +2,31 @@ package cmd
 
 import (
 	"fmt"
-	"github.com/stretchr/testify/require"
 	"net"
-	"os"
 	"testing"
 	"time"
+
+	"github.com/ErdemOzgen/blackdagger/internal/test"
+	"github.com/stretchr/testify/require"
 )
 
 func TestServerCommand(t *testing.T) {
-	tmpDir, _, _ := setupTest(t)
-	defer func() {
-		_ = os.RemoveAll(tmpDir)
-	}()
+	t.Run("StartServer", func(t *testing.T) {
+		setup := test.SetupTest(t)
+		defer setup.Cleanup()
 
-	go func() {
-		testRunCommand(t, serverCmd(), cmdTest{
-			args:        []string{"server", fmt.Sprintf("--port=%s", findPort(t))},
-			expectedOut: []string{"server is running"},
-		})
-	}()
+		go func() {
+			testRunCommand(t, serverCmd(), cmdTest{
+				args:        []string{"server", fmt.Sprintf("--port=%s", findPort(t))},
+				expectedOut: []string{"server is running"},
+			})
+		}()
 
-	time.Sleep(time.Millisecond * 500)
+		time.Sleep(time.Millisecond * 500)
+	})
 }
 
+// findPort finds an available port.
 func findPort(t *testing.T) string {
 	t.Helper()
 	ln, err := net.Listen("tcp", ":0")
