@@ -1,27 +1,30 @@
 package cmd
 
 import (
-	"os"
 	"testing"
+
+	"github.com/ErdemOzgen/blackdagger/internal/test"
 )
 
 func TestStartCommand(t *testing.T) {
-	tmpDir, _, _ := setupTest(t)
-	defer func() {
-		_ = os.RemoveAll(tmpDir)
-	}()
+	setup := test.SetupTest(t)
+	defer setup.Cleanup()
 
 	tests := []cmdTest{
 		{
-			args:        []string{"start", testDAGFile("start.yaml")},
+			args:        []string{"start", testDAGFile("success.yaml")},
 			expectedOut: []string{"1 finished"},
 		},
 		{
-			args:        []string{"start", testDAGFile("start_with_params.yaml")},
+			args:        []string{"start", testDAGFile("params.yaml")},
 			expectedOut: []string{"params is p1 and p2"},
 		},
 		{
-			args:        []string{"start", `--params="p3 p4"`, testDAGFile("start_with_params.yaml")},
+			args: []string{
+				"start",
+				`--params="p3 p4"`,
+				testDAGFile("params.yaml"),
+			},
 			expectedOut: []string{"params is p3 and p4"},
 		},
 	}
