@@ -1,34 +1,17 @@
 package cmd
 
 import (
-	"fmt"
-
-	"github.com/ErdemOzgen/blackdagger/internal/config"
 	"github.com/ErdemOzgen/blackdagger/internal/constants"
 	"github.com/spf13/cobra"
 )
-
-var AsciiArt = `
-
-██████╗░██╗░░░░░░█████╗░░█████╗░██╗░░██╗██████╗░░█████╗░░██████╗░░██████╗░███████╗██████╗░
-██╔══██╗██║░░░░░██╔══██╗██╔══██╗██║░██╔╝██╔══██╗██╔══██╗██╔════╝░██╔════╝░██╔════╝██╔══██╗
-██████╦╝██║░░░░░███████║██║░░╚═╝█████═╝░██║░░██║███████║██║░░██╗░██║░░██╗░█████╗░░██████╔╝
-██╔══██╗██║░░░░░██╔══██║██║░░██╗██╔═██╗░██║░░██║██╔══██║██║░░╚██╗██║░░╚██╗██╔══╝░░██╔══██╗
-██████╦╝███████╗██║░░██║╚█████╔╝██║░╚██╗██████╔╝██║░░██║╚██████╔╝╚██████╔╝███████╗██║░░██║
-╚═════╝░╚══════╝╚═╝░░╚═╝░╚════╝░╚═╝░░╚═╝╚═════╝░╚═╝░░╚═╝░╚═════╝░░╚═════╝░╚══════╝╚═╝░░╚═╝              
-`
 
 func versionCmd() *cobra.Command {
 	return &cobra.Command{
 		Use:   "version",
 		Short: "Display the binary version",
 		Long:  `blackdagger version`,
-		PreRun: func(cmd *cobra.Command, args []string) {
-			cobra.CheckErr(config.LoadConfig(homeDir))
-		},
-		Run: func(cmd *cobra.Command, args []string) {
-			fmt.Println(AsciiArt)
-			fmt.Println(constants.Version)
+		Run: func(_ *cobra.Command, _ []string) {
+			println(constants.Version)
 		},
 	}
 }
