@@ -4,330 +4,385 @@ import (
 	"context"
 	"net/http"
 	"net/url"
-	"os"
-	"path"
+	"path/filepath"
 	"syscall"
 	"testing"
 	"time"
 
 	"github.com/ErdemOzgen/blackdagger/internal/agent"
-	"github.com/ErdemOzgen/blackdagger/internal/persistence"
-	"github.com/ErdemOzgen/blackdagger/internal/persistence/client"
+	"github.com/ErdemOzgen/blackdagger/internal/test"
+	"github.com/google/uuid"
 
-	"github.com/ErdemOzgen/blackdagger/internal/config"
 	"github.com/ErdemOzgen/blackdagger/internal/dag"
-	"github.com/ErdemOzgen/blackdagger/internal/engine"
+	"github.com/ErdemOzgen/blackdagger/internal/dag/scheduler"
 	"github.com/ErdemOzgen/blackdagger/internal/persistence/model"
-	"github.com/ErdemOzgen/blackdagger/internal/scheduler"
-	"github.com/ErdemOzgen/blackdagger/internal/utils"
+	"github.com/ErdemOzgen/blackdagger/internal/util"
 	"github.com/stretchr/testify/require"
 )
 
-var testdataDir = path.Join(utils.MustGetwd(), "testdata")
+func TestAgent_Run(t *testing.T) {
+	t.Run("RunDAG", func(t *testing.T) {
+		setup := test.SetupTest(t)
+		defer setup.Cleanup()
 
-func setupTest(t *testing.T) (string, engine.Engine, persistence.DataStoreFactory) {
-	t.Helper()
+		workflow := testLoadDAG(t, "run.yaml")
+		cli := setup.Client()
+		agt := newAgent(setup, genRequestID(), workflow, &agent.Options{})
 
-	tmpDir := utils.MustTempDir("blackdagger_test")
-	_ = os.Setenv("HOME", tmpDir)
-	_ = config.LoadConfig(tmpDir)
-
-	ds := client.NewDataStoreFactory(&config.Config{
-		DataDir: path.Join(tmpDir, ".blackdagger", "data"),
-	})
-
-	e := engine.NewFactory(ds, nil).Create()
-
-	return tmpDir, e, ds
-}
-
-func TestRunDAG(t *testing.T) {
-	tmpDir, e, df := setupTest(t)
-	defer func() {
-		_ = os.RemoveAll(tmpDir)
-	}()
-
-	d := testLoadDAG(t, "run.yaml")
-	a := agent.New(&agent.Config{DAG: d}, e, df)
-
-	status, _ := e.GetLatestStatus(d)
-	require.Equal(t, scheduler.SchedulerStatus_None, status.Status)
-
-	go func() {
-		err := a.Run(context.Background())
+		latestStatus, err := cli.GetLatestStatus(workflow)
 		require.NoError(t, err)
-	}()
-
-	time.Sleep(100 * time.Millisecond)
-
-	require.Eventually(t, func() bool {
-		status, err := e.GetLatestStatus(d)
-		require.NoError(t, err)
-		return status.Status == scheduler.SchedulerStatus_Success
-	}, time.Second*2, time.Millisecond*100)
-
-	// check deletion of expired history files
-	d.HistRetentionDays = 0
-	a = agent.New(&agent.Config{DAG: d}, e, df)
-	err := a.Run(context.Background())
-	require.NoError(t, err)
-	statusList := e.GetRecentHistory(d, 100)
-	require.Equal(t, 1, len(statusList))
-}
-
-func TestCheckRunning(t *testing.T) {
-	tmpDir, e, df := setupTest(t)
-	defer func() {
-		_ = os.RemoveAll(tmpDir)
-	}()
-
-	d := testLoadDAG(t, "is_running.yaml")
-	a := agent.New(&agent.Config{DAG: d}, e, df)
-
-	go func() {
-		_ = a.Run(context.Background())
-	}()
-
-	time.Sleep(time.Millisecond * 30)
-
-	status := a.Status()
-	require.NotNil(t, status)
-	require.Equal(t, status.Status, scheduler.SchedulerStatus_Running)
-
-	a = agent.New(&agent.Config{DAG: d}, e, df)
-	err := a.Run(context.Background())
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "is already running")
-}
-
-func TestDryRun(t *testing.T) {
-	tmpDir, e, df := setupTest(t)
-	defer func() {
-		_ = os.RemoveAll(tmpDir)
-	}()
-
-	d := testLoadDAG(t, "dry.yaml")
-	a := agent.New(&agent.Config{DAG: d, Dry: true}, e, df)
-
-	err := a.Run(context.Background())
-	require.NoError(t, err)
-
-	status := a.Status()
-	require.NoError(t, err)
-
-	require.Equal(t, scheduler.SchedulerStatus_Success, status.Status)
-}
-
-func TestCancelDAG(t *testing.T) {
-	tmpDir, e, df := setupTest(t)
-	defer func() {
-		_ = os.RemoveAll(tmpDir)
-	}()
-
-	for _, abort := range []func(*agent.Agent){
-		func(a *agent.Agent) { a.Signal(syscall.SIGTERM) },
-	} {
-		d := testLoadDAG(t, "sleep.yaml")
-		a := agent.New(&agent.Config{DAG: d}, e, df)
+		require.Equal(t, scheduler.StatusNone, latestStatus.Status)
 
 		go func() {
-			_ = a.Run(context.Background())
+			err := agt.Run(context.Background())
+			require.NoError(t, err)
 		}()
 
-		time.Sleep(time.Millisecond * 100)
-		abort(a)
-		time.Sleep(time.Millisecond * 500)
-		status, err := e.GetLatestStatus(d)
+		time.Sleep(100 * time.Millisecond)
+
+		require.Eventually(t, func() bool {
+			status, err := cli.GetLatestStatus(workflow)
+			require.NoError(t, err)
+			return status.Status == scheduler.StatusSuccess
+		}, time.Second*2, time.Millisecond*100)
+	})
+	t.Run("DeleteOldHistory", func(t *testing.T) {
+		setup := test.SetupTest(t)
+		defer setup.Cleanup()
+
+		// Create a history file by running a DAG
+		workflow := testLoadDAG(t, "simple.yaml")
+		cli := setup.Client()
+		agt := newAgent(setup, genRequestID(), workflow, &agent.Options{})
+
+		err := agt.Run(context.Background())
 		require.NoError(t, err)
-		require.Equal(t, scheduler.SchedulerStatus_Cancel, status.Status)
-	}
-}
+		history := cli.GetRecentHistory(workflow, 2)
+		require.Equal(t, 1, len(history))
 
-func TestPreConditionInvalid(t *testing.T) {
-	tmpDir, e, df := setupTest(t)
-	defer func() {
-		_ = os.RemoveAll(tmpDir)
-	}()
-	d := testLoadDAG(t, "multiple_steps.yaml")
+		// Set the retention days to 0 and run the DAG again
+		workflow.HistRetentionDays = 0
+		agt = newAgent(setup, genRequestID(), workflow, &agent.Options{})
+		err = agt.Run(context.Background())
+		require.NoError(t, err)
 
-	d.Preconditions = []*dag.Condition{
-		{
-			Condition: "`echo 1`",
-			Expected:  "0",
-		},
-	}
+		// Check if only the latest history file exists
+		history = cli.GetRecentHistory(workflow, 2)
+		require.Equal(t, 1, len(history))
+	})
+	t.Run("AlreadyRunning", func(t *testing.T) {
+		setup := test.SetupTest(t)
+		defer setup.Cleanup()
 
-	a := agent.New(&agent.Config{DAG: d}, e, df)
-	err := a.Run(context.Background())
-	require.Error(t, err)
+		workflow := testLoadDAG(t, "is_running.yaml")
+		agt := newAgent(setup, genRequestID(), workflow, &agent.Options{})
 
-	status := a.Status()
+		go func() {
+			_ = agt.Run(context.Background())
+		}()
 
-	require.Equal(t, scheduler.SchedulerStatus_Cancel, status.Status)
-	require.Equal(t, scheduler.NodeStatus_None, status.Nodes[0].Status)
-	require.Equal(t, scheduler.NodeStatus_None, status.Nodes[1].Status)
-}
+		time.Sleep(time.Millisecond * 30)
 
-func TestPreConditionValid(t *testing.T) {
-	tmpDir, e, df := setupTest(t)
-	defer func() {
-		_ = os.RemoveAll(tmpDir)
-	}()
-	d := testLoadDAG(t, "with_params.yaml")
+		curStatus := agt.Status()
+		require.NotNil(t, curStatus)
+		require.Equal(t, curStatus.Status, scheduler.StatusRunning)
 
-	d.Preconditions = []*dag.Condition{
-		{
-			Condition: "`echo 1`",
-			Expected:  "1",
-		},
-	}
+		agt = newAgent(setup, genRequestID(), workflow, &agent.Options{})
+		err := agt.Run(context.Background())
+		require.Error(t, err)
+		require.Contains(t, err.Error(), "is already running")
+	})
+	t.Run("PreConditionNotMet", func(t *testing.T) {
+		setup := test.SetupTest(t)
+		defer setup.Cleanup()
 
-	a := agent.New(&agent.Config{DAG: d}, e, df)
-	err := a.Run(context.Background())
-	require.NoError(t, err)
+		workflow := testLoadDAG(t, "multiple_steps.yaml")
 
-	status := a.Status()
-	require.Equal(t, scheduler.SchedulerStatus_Success, status.Status)
-	for _, s := range status.Nodes {
-		require.Equal(t, scheduler.NodeStatus_Success, s.Status)
-	}
-}
+		// Precondition is not met
+		workflow.Preconditions = []dag.Condition{{Condition: "`echo 1`", Expected: "0"}}
 
-func TestStartError(t *testing.T) {
-	tmpDir, e, df := setupTest(t)
-	defer func() {
-		_ = os.RemoveAll(tmpDir)
-	}()
-	d := testLoadDAG(t, "error.yaml")
+		agt := newAgent(setup, genRequestID(), workflow, &agent.Options{})
+		err := agt.Run(context.Background())
+		require.Error(t, err)
 
-	a := agent.New(&agent.Config{DAG: d}, e, df)
-	err := a.Run(context.Background())
-	require.Error(t, err)
+		// Check if all nodes are not executed
+		status := agt.Status()
+		require.Equal(t, scheduler.StatusCancel, status.Status)
+		require.Equal(t, scheduler.NodeStatusNone, status.Nodes[0].Status)
+		require.Equal(t, scheduler.NodeStatusNone, status.Nodes[1].Status)
+	})
+	t.Run("FinishWithError", func(t *testing.T) {
+		setup := test.SetupTest(t)
+		defer setup.Cleanup()
 
-	status := a.Status()
-	require.Equal(t, scheduler.SchedulerStatus_Error, status.Status)
-}
+		// Run a DAG that fails
+		errDAG := testLoadDAG(t, "error.yaml")
+		agt := newAgent(setup, genRequestID(), errDAG, &agent.Options{})
+		err := agt.Run(context.Background())
+		require.Error(t, err)
 
-func TestOnExit(t *testing.T) {
-	tmpDir, e, df := setupTest(t)
-	defer func() {
-		_ = os.RemoveAll(tmpDir)
-	}()
+		// Check if the status is saved correctly
+		require.Equal(t, scheduler.StatusError, agt.Status().Status)
+	})
+	t.Run("FinishWithTimeout", func(t *testing.T) {
+		setup := test.SetupTest(t)
+		defer setup.Cleanup()
 
-	d := testLoadDAG(t, "on_exit.yaml")
-	a := agent.New(&agent.Config{DAG: d}, e, df)
-	err := a.Run(context.Background())
-	require.NoError(t, err)
+		// Run a DAG that timeout
+		timeoutDAG := testLoadDAG(t, "timeout.yaml")
+		agt := newAgent(setup, genRequestID(), timeoutDAG, &agent.Options{})
+		ctx := context.Background()
+		err := agt.Run(ctx)
+		require.Error(t, err)
 
-	status := a.Status()
-	require.Equal(t, scheduler.SchedulerStatus_Success, status.Status)
-	for _, s := range status.Nodes {
-		require.Equal(t, scheduler.NodeStatus_Success, s.Status)
-	}
-	require.Equal(t, scheduler.NodeStatus_Success, status.OnExit.Status)
-}
+		// Check if the status is saved correctly
+		require.Equal(t, scheduler.StatusError, agt.Status().Status)
+	})
+	t.Run("ReceiveSignal", func(t *testing.T) {
+		setup := test.SetupTest(t)
+		defer setup.Cleanup()
 
-func TestRetry(t *testing.T) {
-	tmpDir, e, df := setupTest(t)
-	defer func() {
-		_ = os.RemoveAll(tmpDir)
-	}()
+		abortFunc := func(a *agent.Agent) { a.Signal(syscall.SIGTERM) }
 
-	d := testLoadDAG(t, "retry.yaml")
+		workflow := testLoadDAG(t, "sleep.yaml")
+		cli := setup.Client()
+		agt := newAgent(setup, genRequestID(), workflow, &agent.Options{})
 
-	a := agent.New(&agent.Config{DAG: d}, e, df)
-	err := a.Run(context.Background())
-	require.Error(t, err)
+		go func() {
+			_ = agt.Run(context.Background())
+		}()
 
-	status := a.Status()
-	require.Equal(t, scheduler.SchedulerStatus_Error, status.Status)
+		// wait for the DAG to start
+		require.Eventually(t, func() bool {
+			status, err := cli.GetLatestStatus(workflow)
+			require.NoError(t, err)
+			return status.Status == scheduler.StatusRunning
+		}, time.Second*1, time.Millisecond*100)
 
-	for _, n := range status.Nodes {
-		n.CmdWithArgs = "true"
-	}
+		// send a signal to cancel the DAG
+		abortFunc(agt)
 
-	a = agent.New(&agent.Config{DAG: d, RetryTarget: status}, e, df)
-	err = a.Run(context.Background())
-	require.NoError(t, err)
+		require.Eventually(t, func() bool {
+			status, err := cli.GetLatestStatus(workflow)
+			require.NoError(t, err)
+			return status.Status == scheduler.StatusCancel
+		}, time.Second*1, time.Millisecond*100)
+	})
+	t.Run("ExitHandler", func(t *testing.T) {
+		setup := test.SetupTest(t)
+		defer setup.Cleanup()
 
-	status = a.Status()
-	require.Equal(t, scheduler.SchedulerStatus_Success, status.Status)
+		workflow := testLoadDAG(t, "on_exit.yaml")
+		agt := newAgent(setup, genRequestID(), workflow, &agent.Options{})
+		err := agt.Run(context.Background())
+		require.NoError(t, err)
 
-	for _, n := range status.Nodes {
-		if n.Status != scheduler.NodeStatus_Success &&
-			n.Status != scheduler.NodeStatus_Skipped && n.Status != scheduler.NodeStatus_ValidSkip {
-			t.Errorf("invalid status: %s", n.Status.String())
+		// Check if the DAG is executed successfully
+		status := agt.Status()
+		require.Equal(t, scheduler.StatusSuccess, status.Status)
+		for _, s := range status.Nodes {
+			require.Equal(t, scheduler.NodeStatusSuccess, s.Status)
 		}
-	}
+
+		// Check if the exit handler is executed
+		require.Equal(t, scheduler.NodeStatusSuccess, status.OnExit.Status)
+	})
 }
 
-func TestHandleHTTP(t *testing.T) {
-	tmpDir, e, df := setupTest(t)
-	defer func() {
-		_ = os.RemoveAll(tmpDir)
-	}()
+func TestAgent_DryRun(t *testing.T) {
+	t.Parallel()
+	t.Run("DryRun", func(t *testing.T) {
+		setup := test.SetupTest(t)
+		defer setup.Cleanup()
 
-	d := testLoadDAG(t, "handle_http.yaml")
-	a := agent.New(&agent.Config{DAG: d}, e, df)
+		workflow := testLoadDAG(t, "dry.yaml")
+		agt := newAgent(setup, genRequestID(), workflow, &agent.Options{
+			Dry: true,
+		})
 
-	go func() {
-		err := a.Run(context.Background())
+		err := agt.Run(context.Background())
 		require.NoError(t, err)
-	}()
 
-	<-time.After(time.Millisecond * 50)
+		curStatus := agt.Status()
+		require.NoError(t, err)
+		require.Equal(t, scheduler.StatusSuccess, curStatus.Status)
 
-	var mockResponseWriter = mockResponseWriter{}
-
-	// status
-	req := &http.Request{
-		Method: "GET",
-		URL: &url.URL{
-			Path: "/status",
-		},
-	}
-
-	a.HandleHTTP(&mockResponseWriter, req)
-	require.Equal(t, http.StatusOK, mockResponseWriter.status)
-
-	status, err := model.StatusFromJson(mockResponseWriter.body)
-	require.NoError(t, err)
-	require.Equal(t, scheduler.SchedulerStatus_Running, status.Status)
-
-	// invalid path
-	req = &http.Request{
-		Method: "GET",
-		URL: &url.URL{
-			Path: "/invalid-path",
-		},
-	}
-	a.HandleHTTP(&mockResponseWriter, req)
-	require.Equal(t, http.StatusNotFound, mockResponseWriter.status)
-
-	// cancel
-	req = &http.Request{
-		Method: "POST",
-		URL: &url.URL{
-			Path: "/stop",
-		},
-	}
-	a.HandleHTTP(&mockResponseWriter, req)
-	require.Equal(t, http.StatusOK, mockResponseWriter.status)
-	require.Equal(t, "OK", mockResponseWriter.body)
-
-	<-time.After(time.Millisecond * 50)
-
-	status = a.Status()
-	require.Equal(t, status.Status, scheduler.SchedulerStatus_Cancel)
+		// Check if the status is not saved
+		cli := setup.Client()
+		history := cli.GetRecentHistory(workflow, 1)
+		require.Equal(t, 0, len(history))
+	})
 }
+
+func TestAgent_Retry(t *testing.T) {
+	t.Parallel()
+	t.Run("RetryDAG", func(t *testing.T) {
+		setup := test.SetupTest(t)
+		defer setup.Cleanup()
+
+		// retry.yaml has a DAG that fails
+		workflow := testLoadDAG(t, "retry.yaml")
+
+		agt := newAgent(setup, genRequestID(), workflow, &agent.Options{})
+		err := agt.Run(context.Background())
+		require.Error(t, err)
+
+		// Check if the DAG failed
+		status := agt.Status()
+		require.Equal(t, scheduler.StatusError, status.Status)
+
+		// Modify the DAG to make it successful
+		for _, node := range status.Nodes {
+			node.Step.CmdWithArgs = "true"
+		}
+
+		// Retry the DAG and check if it is successful
+		agt = newAgent(setup, genRequestID(), workflow, &agent.Options{
+			RetryTarget: status,
+		})
+		err = agt.Run(context.Background())
+		require.NoError(t, err)
+
+		status = agt.Status()
+		require.Equal(t, scheduler.StatusSuccess, status.Status)
+
+		for _, node := range status.Nodes {
+			if node.Status != scheduler.NodeStatusSuccess &&
+				node.Status != scheduler.NodeStatusSkipped {
+				t.Errorf("invalid status: %s", node.Status.String())
+			}
+		}
+	})
+}
+
+func TestAgent_HandleHTTP(t *testing.T) {
+	t.Parallel()
+	t.Run("HTTP_Valid", func(t *testing.T) {
+		setup := test.SetupTest(t)
+		defer setup.Cleanup()
+
+		// Start a long-running DAG
+		workflow := testLoadDAG(t, "handle_http.yaml")
+		agt := newAgent(setup, genRequestID(), workflow, &agent.Options{})
+		go func() {
+			err := agt.Run(context.Background())
+			require.NoError(t, err)
+		}()
+
+		// Wait for the DAG to start
+		cli := setup.Client()
+		require.Eventually(t, func() bool {
+			status, _ := cli.GetLatestStatus(workflow)
+			// require.NoError(t, err)
+			return status.Status == scheduler.StatusRunning
+		}, time.Second*2, time.Millisecond*100)
+
+		// Get the status of the DAG
+		var mockResponseWriter = mockResponseWriter{}
+		agt.HandleHTTP(&mockResponseWriter, &http.Request{
+			Method: "GET", URL: &url.URL{Path: "/status"},
+		})
+		require.Equal(t, http.StatusOK, mockResponseWriter.status)
+
+		// Check if the status is returned correctly
+		status, err := model.StatusFromJSON(mockResponseWriter.body)
+		require.NoError(t, err)
+		require.Equal(t, scheduler.StatusRunning, status.Status)
+
+		// Stop the DAG
+		agt.Signal(syscall.SIGTERM)
+		require.Eventually(t, func() bool {
+			status, err := cli.GetLatestStatus(workflow)
+			require.NoError(t, err)
+			return status.Status == scheduler.StatusCancel
+		}, time.Second*2, time.Millisecond*100)
+
+	})
+	t.Run("HTTP_InvalidRequest", func(t *testing.T) {
+		setup := test.SetupTest(t)
+		defer setup.Cleanup()
+
+		// Start a long-running DAG
+		workflow := testLoadDAG(t, "handle_http2.yaml")
+		agt := newAgent(setup, genRequestID(), workflow, &agent.Options{})
+
+		go func() {
+			err := agt.Run(context.Background())
+			require.NoError(t, err)
+		}()
+
+		// Wait for the DAG to start
+		cli := setup.Client()
+		require.Eventually(t, func() bool {
+			status, err := cli.GetLatestStatus(workflow)
+			require.NoError(t, err)
+			return status.Status == scheduler.StatusRunning
+		}, time.Second*2, time.Millisecond*100)
+
+		var mockResponseWriter = mockResponseWriter{}
+
+		// Request with an invalid path
+		agt.HandleHTTP(&mockResponseWriter, &http.Request{
+			Method: "GET",
+			URL:    &url.URL{Path: "/invalid-path"},
+		})
+		require.Equal(t, http.StatusNotFound, mockResponseWriter.status)
+
+		// Stop the DAG
+		agt.Signal(syscall.SIGTERM)
+		require.Eventually(t, func() bool {
+			status, err := cli.GetLatestStatus(workflow)
+			require.NoError(t, err)
+			return status.Status == scheduler.StatusCancel
+		}, time.Second*2, time.Millisecond*100)
+	})
+	t.Run("HTTP_HandleCancel", func(t *testing.T) {
+		setup := test.SetupTest(t)
+		defer setup.Cleanup()
+
+		// Start a long-running DAG
+		workflow := testLoadDAG(t, "handle_http3.yaml")
+		agt := newAgent(setup, genRequestID(), workflow, &agent.Options{})
+
+		go func() {
+			err := agt.Run(context.Background())
+			require.NoError(t, err)
+		}()
+
+		// Wait for the DAG to start
+		cli := setup.Client()
+		require.Eventually(t, func() bool {
+			status, err := cli.GetLatestStatus(workflow)
+			require.NoError(t, err)
+			return status.Status == scheduler.StatusRunning
+		}, time.Second*2, time.Millisecond*100)
+
+		// Cancel the DAG
+		var mockResponseWriter = mockResponseWriter{}
+		agt.HandleHTTP(&mockResponseWriter, &http.Request{
+			Method: "POST",
+			URL:    &url.URL{Path: "/stop"},
+		})
+		require.Equal(t, http.StatusOK, mockResponseWriter.status)
+		require.Equal(t, "OK", mockResponseWriter.body)
+
+		// Wait for the DAG to stop
+		require.Eventually(t, func() bool {
+			status, err := cli.GetLatestStatus(workflow)
+			require.NoError(t, err)
+			return status.Status == scheduler.StatusCancel
+		}, time.Second*3, time.Millisecond*100)
+	})
+}
+
+// Assert that mockResponseWriter implements http.ResponseWriter
+var _ http.ResponseWriter = (*mockResponseWriter)(nil)
 
 type mockResponseWriter struct {
 	status int
 	body   string
 	header *http.Header
 }
-
-var _ http.ResponseWriter = (*mockResponseWriter)(nil)
 
 func (h *mockResponseWriter) Header() http.Header {
 	if h.header == nil {
@@ -338,17 +393,44 @@ func (h *mockResponseWriter) Header() http.Header {
 
 func (h *mockResponseWriter) Write(body []byte) (int, error) {
 	h.body = string(body)
-	return 0, nil
+	return len([]byte(h.body)), nil
 }
 
 func (h *mockResponseWriter) WriteHeader(statusCode int) {
 	h.status = statusCode
 }
 
+// testLoadDAG load the specified DAG file for testing
+// without base config or parameters.
 func testLoadDAG(t *testing.T, name string) *dag.DAG {
-	file := path.Join(testdataDir, name)
-	cl := &dag.Loader{}
-	d, err := cl.Load(file, "")
+	file := filepath.Join(util.MustGetwd(), "testdata", name)
+	workflow, err := dag.Load("", file, "")
 	require.NoError(t, err)
-	return d
+	return workflow
+}
+
+func genRequestID() string {
+	id, err := uuid.NewRandom()
+	if err != nil {
+		panic(err)
+	}
+	return id.String()
+}
+
+func newAgent(
+	setup test.Setup,
+	requestID string,
+	workflow *dag.DAG,
+	opts *agent.Options,
+) *agent.Agent {
+	return agent.New(
+		requestID,
+		workflow,
+		test.NewLogger(),
+		setup.Config.LogDir,
+		"",
+		setup.Client(),
+		setup.DataStore(),
+		opts,
+	)
 }
