@@ -2,10 +2,11 @@ package local
 
 import (
 	"fmt"
-	"github.com/ErdemOzgen/blackdagger/internal/persistence"
-	"github.com/ErdemOzgen/blackdagger/internal/persistence/local/storage"
 	"regexp"
 	"strings"
+
+	"github.com/ErdemOzgen/blackdagger/internal/persistence"
+	"github.com/ErdemOzgen/blackdagger/internal/persistence/local/storage"
 )
 
 type flagStoreImpl struct {
@@ -37,8 +38,12 @@ func fileName(id string) string {
 
 // https://github.com/sindresorhus/filename-reserved-regex/blob/master/index.js
 var (
-	filenameReservedRegex             = regexp.MustCompile(`[<>:"/\\|?*\x00-\x1F]`)
-	filenameReservedWindowsNamesRegex = regexp.MustCompile(`(?i)^(con|prn|aux|nul|com[0-9]|lpt[0-9])$`)
+	filenameReservedRegex = regexp.MustCompile(
+		`[<>:"/\\|?*\x00-\x1F]`,
+	)
+	filenameReservedWindowsNamesRegex = regexp.MustCompile(
+		`(?i)^(con|prn|aux|nul|com[0-9]|lpt[0-9])$`,
+	)
 )
 
 func normalizeFilename(str, replacement string) string {
