@@ -10,9 +10,14 @@ type Storage struct {
 	Dir string
 }
 
+var (
+	// TODO: use 0600 // nolint: gosec
+	defaultPermission os.FileMode = 0744
+)
+
 // NewStorage creates a new storage.
 func NewStorage(dir string) *Storage {
-	_ = os.MkdirAll(dir, 0755)
+	_ = os.MkdirAll(dir, defaultPermission)
 
 	return &Storage{
 		Dir: dir,
@@ -21,7 +26,7 @@ func NewStorage(dir string) *Storage {
 
 // Create creates the given file.
 func (s *Storage) Create(file string) error {
-	return os.WriteFile(path.Join(s.Dir, file), []byte{}, 0644)
+	return os.WriteFile(path.Join(s.Dir, file), []byte{}, defaultPermission)
 }
 
 // Exists returns true if the given file exists.
