@@ -4,56 +4,16 @@ import (
 	"encoding/json"
 	"fmt"
 	"strings"
+	"sync"
 	"time"
 
 	"github.com/ErdemOzgen/blackdagger/internal/dag"
-	"github.com/ErdemOzgen/blackdagger/internal/scheduler"
-	"github.com/ErdemOzgen/blackdagger/internal/utils"
+	"github.com/ErdemOzgen/blackdagger/internal/dag/scheduler"
+	"github.com/ErdemOzgen/blackdagger/internal/util"
 )
 
-type StatusResponse struct {
-	Status *Status `json:"status"`
-}
-
-type Pid int
-
-const PidNotRunning Pid = -1
-
-func (p Pid) String() string {
-	if p == PidNotRunning {
-		return ""
-	}
-	return fmt.Sprintf("%d", p)
-}
-
-func (p Pid) IsRunning() bool {
-	return p != PidNotRunning
-}
-
-type Status struct {
-	RequestId  string                    `json:"RequestId"`
-	Name       string                    `json:"Name"`
-	Status     scheduler.SchedulerStatus `json:"Status"`
-	StatusText string                    `json:"StatusText"`
-	Pid        Pid                       `json:"Pid"`
-	Nodes      []*Node                   `json:"Nodes"`
-	OnExit     *Node                     `json:"OnExit"`
-	OnSuccess  *Node                     `json:"OnSuccess"`
-	OnFailure  *Node                     `json:"OnFailure"`
-	OnCancel   *Node                     `json:"OnCancel"`
-	StartedAt  string                    `json:"StartedAt"`
-	FinishedAt string                    `json:"FinishedAt"`
-	Log        string                    `json:"Log"`
-	Params     string                    `json:"Params"`
-}
-
-type StatusFile struct {
-	File   string
-	Status *Status
-}
-
-func StatusFromJson(s string) (*Status, error) {
-	status := &Status{}
+func StatusFromJSON(s string) (*Status, error) {
+	status := new(Status)
 	err := json.Unmarshal([]byte(s), status)
 	if err != nil {
 		return nil, err
@@ -61,63 +21,124 @@ func StatusFromJson(s string) (*Status, error) {
 	return status, err
 }
 
-func NewStatusDefault(d *dag.DAG) *Status {
-	return NewStatus(d, nil, scheduler.SchedulerStatus_None, int(PidNotRunning), nil, nil)
+func FromNodesOrSteps(nodes []scheduler.NodeData, steps []dag.Step) []*Node {
+	if len(nodes) != 0 {
+		return FromNodes(nodes)
+	}
+	return FromSteps(steps)
+}
+
+type StatusFile struct {
+	File   string
+	Status *Status
+}
+
+type StatusResponse struct {
+	Status *Status `json:"status"`
+}
+
+type Status struct {
+	RequestID  string           `json:"RequestId"`
+	Name       string           `json:"Name"`
+	Status     scheduler.Status `json:"Status"`
+	StatusText string           `json:"StatusText"`
+	PID        PID              `json:"Pid"`
+	Nodes      []*Node          `json:"Nodes"`
+	OnExit     *Node            `json:"OnExit"`
+	OnSuccess  *Node            `json:"OnSuccess"`
+	OnFailure  *Node            `json:"OnFailure"`
+	OnCancel   *Node            `json:"OnCancel"`
+	StartedAt  string           `json:"StartedAt"`
+	FinishedAt string           `json:"FinishedAt"`
+	Log        string           `json:"Log"`
+	Params     string           `json:"Params"`
+	mu         sync.RWMutex
+}
+
+func NewStatusDefault(workflow *dag.DAG) *Status {
+	return NewStatus(
+		workflow, nil, scheduler.StatusNone, int(pidNotRunning), nil, nil,
+	)
 }
 
 func NewStatus(
-	d *dag.DAG,
-	nodes []*scheduler.Node,
-	status scheduler.SchedulerStatus,
+	workflow *dag.DAG,
+	nodes []scheduler.NodeData,
+	status scheduler.Status,
 	pid int,
-	startTime, endTIme *time.Time,
+	startTime, endTime *time.Time,
 ) *Status {
-	finish, start := time.Time{}, time.Time{}
-	if startTime != nil {
-		start = *startTime
-	}
-	if endTIme != nil {
-		finish = *endTIme
-	}
-	var models []*Node
-	if len(nodes) != 0 {
-		models = FromNodes(nodes)
-	} else {
-		models = FromSteps(d.Steps)
-	}
-	var onExit, onSuccess, onFailure, onCancel *Node
-	onExit = fromStepWithDefValues(d.HandlerOn.Exit)
-	onSuccess = fromStepWithDefValues(d.HandlerOn.Success)
-	onFailure = fromStepWithDefValues(d.HandlerOn.Failure)
-	onCancel = fromStepWithDefValues(d.HandlerOn.Cancel)
-	return &Status{
-		RequestId:  "",
-		Name:       d.Name,
+	statusObj := &Status{
+		Name:       workflow.Name,
 		Status:     status,
 		StatusText: status.String(),
-		Pid:        Pid(pid),
-		Nodes:      models,
-		OnExit:     onExit,
-		OnSuccess:  onSuccess,
-		OnFailure:  onFailure,
-		OnCancel:   onCancel,
-		StartedAt:  utils.FormatTime(start),
-		FinishedAt: utils.FormatTime(finish),
-		Params:     strings.Join(d.Params, " "),
+		PID:        PID(pid),
+		Nodes:      FromNodesOrSteps(nodes, workflow.Steps),
+		OnExit:     nodeOrNil(workflow.HandlerOn.Exit),
+		OnSuccess:  nodeOrNil(workflow.HandlerOn.Success),
+		OnFailure:  nodeOrNil(workflow.HandlerOn.Failure),
+		OnCancel:   nodeOrNil(workflow.HandlerOn.Cancel),
+		Params:     Params(workflow.Params),
+	}
+	if startTime != nil {
+		statusObj.StartedAt = util.FormatTime(*startTime)
+	}
+	if endTime != nil {
+		statusObj.FinishedAt = util.FormatTime(*endTime)
+	}
+	return statusObj
+}
+
+func (st *Status) CorrectRunningStatus() {
+	if st.Status == scheduler.StatusRunning {
+		st.Status = scheduler.StatusError
+		st.StatusText = st.Status.String()
 	}
 }
 
-func (sts *Status) CorrectRunningStatus() {
-	if sts.Status == scheduler.SchedulerStatus_Running {
-		sts.Status = scheduler.SchedulerStatus_Error
-		sts.StatusText = sts.Status.String()
-	}
-}
-
-func (sts *Status) ToJson() ([]byte, error) {
-	js, err := json.Marshal(sts)
+func (st *Status) ToJSON() ([]byte, error) {
+	st.mu.RLock()
+	defer st.mu.RUnlock()
+	js, err := json.Marshal(st)
 	if err != nil {
 		return []byte{}, err
 	}
 	return js, nil
+}
+
+func FormatTime(val time.Time) string {
+	if val.IsZero() {
+		return ""
+	}
+	return util.FormatTime(val)
+}
+
+func Time(t time.Time) *time.Time {
+	return &t
+}
+
+func Params(params []string) string {
+	return strings.Join(params, " ")
+}
+
+type PID int
+
+const pidNotRunning PID = -1
+
+func (p PID) String() string {
+	if p == pidNotRunning {
+		return ""
+	}
+	return fmt.Sprintf("%d", p)
+}
+
+func (p PID) IsRunning() bool {
+	return p != pidNotRunning
+}
+
+func nodeOrNil(s *dag.Step) *Node {
+	if s == nil {
+		return nil
+	}
+	return NewNode(*s)
 }
