@@ -1,14 +1,14 @@
-package sock
+package sock_test
 
 import (
 	"errors"
-	"fmt"
-	"net"
 	"net/http"
 	"os"
 	"testing"
 	"time"
 
+	"github.com/ErdemOzgen/blackdagger/internal/sock"
+	"github.com/ErdemOzgen/blackdagger/internal/test"
 	"github.com/stretchr/testify/require"
 )
 
@@ -19,7 +19,7 @@ func TestDialFail(t *testing.T) {
 		_ = os.Remove(f.Name())
 	}()
 
-	client := Client{Addr: f.Name()}
+	client := sock.NewClient(f.Name())
 	_, err = client.Request("GET", "/status")
 	require.Error(t, err)
 }
@@ -31,41 +31,26 @@ func TestDialTimeout(t *testing.T) {
 		_ = os.Remove(f.Name())
 	}()
 
-	s, err := NewServer(
-		&Config{
-			Addr: f.Name(),
-			HandlerFunc: func(w http.ResponseWriter, r *http.Request) {
-				time.Sleep(time.Second * 3100)
-				w.WriteHeader(http.StatusOK)
-				_, _ = w.Write([]byte("OK"))
-			},
-		})
+	srv, err := sock.NewServer(
+		f.Name(),
+		func(w http.ResponseWriter, _ *http.Request) {
+			time.Sleep(time.Second * 3100)
+			w.WriteHeader(http.StatusOK)
+			_, _ = w.Write([]byte("OK"))
+		},
+		test.NewLogger(),
+	)
 	require.NoError(t, err)
 
 	go func() {
-		_ = s.Serve(nil)
+		_ = srv.Serve(nil)
 	}()
 
 	time.Sleep(time.Millisecond * 500)
 
 	require.NoError(t, err)
-	client := Client{Addr: f.Name()}
+	client := sock.NewClient(f.Name())
 	_, err = client.Request("GET", "/status")
 	require.Error(t, err)
-	require.True(t, errors.Is(err, ErrTimeout))
+	require.True(t, errors.Is(err, sock.ErrTimeout))
 }
-
-func TestProcErr(t *testing.T) {
-	e := procError("test", fmt.Errorf("error"))
-	require.Contains(t, e.Error(), "test failed")
-
-	e = procError("test", errTimeout)
-	require.Contains(t, e.Error(), "test timeout")
-}
-
-type testTimeout struct{ error }
-
-var errTimeout net.Error = &testTimeout{error: fmt.Errorf("timeout")}
-
-func (t *testTimeout) Timeout() bool   { return true }
-func (t *testTimeout) Temporary() bool { return false }
