@@ -4,47 +4,51 @@ import (
 	"bufio"
 	"errors"
 	"io"
-	"log"
 	"net"
 	"net/http"
 	"os"
 	"strings"
+	"sync"
+	"sync/atomic"
 
-	"github.com/ErdemOzgen/blackdagger/internal/utils"
+	"github.com/ErdemOzgen/blackdagger/internal/logger"
+)
+
+var ErrServerRequestedShutdown = errors.New(
+	"socket frontend is requested to shutdown",
 )
 
 // Server is a unix socket frontend that passes http requests to HandlerFunc.
 type Server struct {
-	*Config
-	listener net.Listener
-	quit     bool
+	addr        string
+	handlerFunc HTTPHandlerFunc
+	listener    net.Listener
+	quit        atomic.Bool
+	mu          sync.Mutex
+	logger      logger.Logger
 }
 
-type Config struct {
-	Addr        string
-	HandlerFunc HttpHandlerFunc
-}
-
-// HttpHandlerFunc is a function that handles HTTP requests.
-type HttpHandlerFunc func(w http.ResponseWriter, r *http.Request)
+// HTTPHandlerFunc is a function that handles HTTP requests.
+type HTTPHandlerFunc func(w http.ResponseWriter, r *http.Request)
 
 // NewServer creates a new unix socket frontend.
-func NewServer(c *Config) (*Server, error) {
+func NewServer(
+	addr string,
+	handlerFunc HTTPHandlerFunc,
+	lg logger.Logger,
+) (*Server, error) {
 	return &Server{
-		Config: c,
-		quit:   false,
+		addr:        addr,
+		handlerFunc: handlerFunc,
+		logger:      lg,
 	}, nil
 }
 
-var (
-	ErrServerRequestedShutdown = errors.New("socket frontend is requested to shutdown")
-)
-
 // Serve starts listening and serving requests.
-func (svr *Server) Serve(listen chan error) error {
-	os.Remove(svr.Addr)
+func (srv *Server) Serve(listen chan error) error {
+	_ = os.Remove(srv.addr)
 	var err error
-	svr.listener, err = net.Listen("unix", svr.Addr)
+	srv.listener, err = net.Listen("unix", srv.addr)
 	if err != nil {
 		if listen != nil {
 			listen <- err
@@ -54,41 +58,49 @@ func (svr *Server) Serve(listen chan error) error {
 	if listen != nil {
 		listen <- err
 	}
-	log.Printf("frontend is running at \"%v\"\n", svr.Addr)
+	srv.logger.Debug("Unix socket is listening", "addr", srv.addr)
+
 	defer func() {
-		_ = svr.Shutdown()
-		_ = os.Remove(svr.Addr)
+		_ = srv.Shutdown()
+		_ = os.Remove(srv.addr)
 	}()
 	for {
-		conn, err := svr.listener.Accept()
-		if svr.quit {
+		conn, err := srv.listener.Accept()
+		if srv.quit.Load() {
 			return ErrServerRequestedShutdown
 		}
 		if err == nil {
 			go func() {
 				request, err := http.ReadRequest(bufio.NewReader(conn))
-				utils.LogErr("read request", err)
-				if err == nil {
-					svr.HandlerFunc(newHttpResponseWriter(&conn), request)
+				if err != nil {
+					srv.logger.Error("read request", "error", err)
+				} else {
+					srv.handlerFunc(newHTTPResponseWriter(&conn), request)
 				}
-				conn.Close()
+				_ = conn.Close()
 			}()
 		}
 	}
 }
 
 // Shutdown stops the frontend.
-func (svr *Server) Shutdown() error {
-	if !svr.quit {
-		svr.quit = true
-		if svr.listener != nil {
-			err := svr.listener.Close()
-			utils.LogErr("close listener", err)
+func (srv *Server) Shutdown() error {
+	srv.mu.Lock()
+	defer srv.mu.Unlock()
+	if !srv.quit.Load() {
+		srv.quit.Store(true)
+		if srv.listener != nil {
+			err := srv.listener.Close()
+			if err != nil && !errors.Is(err, os.ErrClosed) {
+				srv.logger.Error("close listener", "error", err)
+			}
 			return err
 		}
 	}
 	return nil
 }
+
+var _ http.ResponseWriter = (*httpResponseWriter)(nil)
 
 type httpResponseWriter struct {
 	conn       *net.Conn
@@ -96,9 +108,7 @@ type httpResponseWriter struct {
 	statusCode int
 }
 
-var _ http.ResponseWriter = (*httpResponseWriter)(nil)
-
-func newHttpResponseWriter(conn *net.Conn) http.ResponseWriter {
+func newHTTPResponseWriter(conn *net.Conn) http.ResponseWriter {
 	return &httpResponseWriter{
 		conn:       conn,
 		header:     make(http.Header),
