@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import {
   flexRender,
   useReactTable,
@@ -27,7 +27,7 @@ import {
   TableRow,
   TextField,
 } from '@mui/material';
-import { Link, useSearchParams } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 import {
   getFirstTag,
   getStatus,
@@ -53,6 +53,10 @@ type Props = {
   DAGs: DAGItem[];
   group: string;
   refreshFn: () => void;
+  searchText: string;
+  handleSearchTextChange: (searchText: string) => void;
+  searchTag: string;
+  handleSearchTagChange: (tag: string) => void;
 };
 
 type DAGRow = DAGItem & { subRows?: DAGItem[] };
@@ -369,11 +373,14 @@ const defaultColumns = [
       if (data.Type == DAGDataType.Group) {
         return null;
       }
+
+      const name = data.DAGStatus.File.replace(/.yaml$/, '');
+
       return (
         <DAGActions
           dag={data.DAGStatus.DAG}
           status={data.DAGStatus.Status}
-          name={data.DAGStatus.DAG.Name}
+          name={name}
           label={false}
           refresh={props.table.options.meta?.refreshFn}
         />
@@ -382,7 +389,7 @@ const defaultColumns = [
   }),
 ];
 
-function DAGTable({ DAGs = [], group = '', refreshFn }: Props) {
+function DAGTable({ DAGs = [], group = '', refreshFn, searchText, handleSearchTextChange, searchTag, handleSearchTagChange }: Props) {
   const [columns] = React.useState<typeof defaultColumns>(() => [
     ...defaultColumns,
   ]);
@@ -396,41 +403,6 @@ function DAGTable({ DAGs = [], group = '', refreshFn }: Props) {
       desc: false,
     },
   ]);
-
-  const [searchParams, setSearchParams] = useSearchParams();
-  useEffect(() => {
-    const searchText = searchParams.get('search');
-    if (searchText) {
-      instance.getColumn('Name')?.setFilterValue(searchText);
-    }
-    const t = searchParams.get('tag');
-    if (t) {
-      instance.getColumn('Tags')?.setFilterValue(t);
-    }
-  }, []);
-
-  const addSearchParam = React.useCallback(
-    (key: string, value: string) => {
-      const ret: { [key: string]: string } = {};
-      searchParams.forEach((v, k) => {
-        if (v && k !== key) {
-          ret[k] = v;
-        }
-      });
-      if (value) {
-        ret[key] = value;
-      }
-      setSearchParams(ret);
-    },
-    [searchParams, setSearchParams]
-  );
-
-  const selectedTag = React.useMemo(() => {
-    return (
-      (columnFilters.find((filter) => filter.id == 'Tags')?.value as string) ||
-      ''
-    );
-  }, [columnFilters]);
 
   const [expanded, setExpanded] = React.useState<ExpandedState>({});
 
@@ -469,23 +441,10 @@ function DAGTable({ DAGs = [], group = '', refreshFn }: Props) {
     ];
   }, [DAGs, group]);
 
-  const tagOptions = React.useMemo(() => {
-    const map: { [key: string]: boolean } = { '': true };
-    DAGs.forEach((data) => {
-      if (data.Type == DAGDataType.DAG) {
-        data.DAGStatus.DAG.Tags?.forEach((tag) => {
-          map[tag] = true;
-        });
-      }
-    });
-    const ret = Object.keys(map).sort();
-    return ret;
-  }, []);
-
-  const instance = useReactTable({
+  const instance = useReactTable<DAGRow>({
     data,
     columns,
-    getSubRows: (row) => row.subRows,
+    getSubRows: (row) =>row.subRows,
     getCoreRowModel: getCoreRowModel(),
     getSortedRowModel: getSortedRowModel(),
     getFilteredRowModel: getFilteredRowModel(),
@@ -525,24 +484,37 @@ function DAGTable({ DAGs = [], group = '', refreshFn }: Props) {
           size="small"
           variant="filled"
           InputProps={{
-            value: instance.getColumn('Name')?.getFilterValue(),
+            value: searchText,
             onChange: (e) => {
               const value = e.target.value || '';
-              addSearchParam('search', value);
-              instance.getColumn('Name')?.setFilterValue(value);
+              handleSearchTextChange(value);
             },
             type: 'search',
           }}
         />
-        <Autocomplete<string>
+        <Autocomplete<string, false, false, true>
           size="small"
           limitTags={1}
-          value={selectedTag}
-          options={tagOptions}
+          value={searchTag}
+          freeSolo
+          options={
+            DAGs.reduce<string[]>((acc, dag) => {
+              if (dag.Type == DAGDataType.DAG) {
+                const tags = dag.DAGStatus.DAG.Tags;
+                if (tags) {
+                  tags.forEach((tag) => {
+                    if (!acc.includes(tag)) {
+                      acc.push(tag);
+                    }
+                  });
+                }
+              }
+              return acc;
+            }, [])
+          }
           onChange={(_, value) => {
             const v = value || '';
-            addSearchParam('tag', v);
-            instance.getColumn('Tags')?.setFilterValue(v);
+            handleSearchTagChange(v);
           }}
           renderInput={(params) => (
             <TextField {...params} variant="filled" label="Search Tag" />
@@ -586,9 +558,9 @@ function DAGTable({ DAGs = [], group = '', refreshFn }: Props) {
                           {header.isPlaceholder
                             ? null
                             : flexRender(
-                                header.column.columnDef.header,
-                                header.getContext()
-                              )}
+                              header.column.columnDef.header,
+                              header.getContext()
+                            )}
                           {{
                             asc: (
                               <ArrowUpward
