@@ -6,13 +6,15 @@ import {
   BarChart,
   Cell,
   LabelList,
+  LabelProps,
   ResponsiveContainer,
   XAxis,
   YAxis,
 } from 'recharts';
 import { statusColorMapping } from '../../consts';
-import { DAGStatus, SchedulerStatus, Status } from '../../models';
-import { WorkflowListItem,WorkflowStatus } from '../../models/api';
+import { DAGStatus } from '../../models';
+import { SchedulerStatus } from '../../models';
+import { WorkflowListItem } from '../../models/api';
 
 type Props = { data: DAGStatus[] | WorkflowListItem[] };
 
@@ -25,29 +27,33 @@ type DataFrame = {
 function DashboardTimechart({ data: input }: Props) {
   const [data, setData] = React.useState<DataFrame[]>([]);
   React.useEffect(() => {
-    const transformedData: DataFrame[] = input.map((dagStatus) => {
-      const status: Status | WorkflowStatus | undefined = dagStatus.Status;
-      if (!status || status.StartedAt === '-' || !status.FinishedAt) {
-        return null;
+    const ret: DataFrame[] = [];
+    const now = moment();
+    const startOfDayUnix = moment().startOf('day').unix();
+    input.forEach((wf) => {
+      const status = wf.Status;
+      const start = status?.StartedAt;
+      if (start && start != '-') {
+        const startUnix = Math.max(moment(start).unix(), startOfDayUnix);
+        const end = status.FinishedAt;
+        let to = now.unix();
+        if (end && end != '-') {
+          to = moment(end).unix();
+        }
+        ret.push({
+          name: status.Name,
+          status: status.Status,
+          values: [startUnix, to],
+        });
       }
-      const startUnix = Math.max(moment(status.StartedAt).unix(), moment().startOf('day').unix());
-      const endUnix = status.FinishedAt && status.FinishedAt !== '-' ? moment(status.FinishedAt).unix() : moment().unix();
-
-      return {
-        name: status.Name,
-        status: status.Status,
-        values: [startUnix, endUnix],
-      };
-    }).filter((item): item is DataFrame => item !== null);
-
-    setData(transformedData.sort((a, b) => a.values[0] - b.values[0]));
+    });
+    const sorted = ret.sort((a, b) => {
+      return a.values[0] < b.values[0] ? -1 : 1;
+    });
+    setData(sorted);
   }, [input]);
-
   const now = moment();
   const shouldScroll = data.length >= 40;
-  // Generating tick values for the XAxis
-  const tickValues = Array.from({ length: 24 }, (_, index) => now.startOf('day').add(index, 'hours').unix());
-
   return (
     <TimelineWrapper shouldScroll={shouldScroll}>
       <ResponsiveContainer
@@ -57,29 +63,42 @@ function DashboardTimechart({ data: input }: Props) {
       >
         <BarChart data={data} layout="vertical">
           <XAxis
+            name="Time"
             tickFormatter={(unixTime) => moment.unix(unixTime).format('HH:mm')}
             type="number"
             dataKey="values"
+            tickCount={96}
             domain={[now.startOf('day').unix(), now.endOf('day').unix()]}
-            ticks={tickValues}
           />
           <YAxis dataKey="name" type="category" hide />
-          <Bar dataKey="values" fill="lightblue" minPointSize={2}>
-            {data.map((entry, index) => (
-              // Use statusColorMapping for setting the cell colors based on status
-              <Cell key={`cell-${index}`} fill={statusColorMapping[entry.status]?.backgroundColor || '#ccc'} />
-            ))}
-            {/* Set the LabelList fill color to bone white */}
-            <LabelList dataKey="name" position="center" fill="#E3DAC9" />
+          <Bar background dataKey="values" fill="lightblue" minPointSize={2}>
+            {data.map((_, index) => {
+              const color = statusColorMapping[data[index].status];
+              return <Cell key={index} fill={color.backgroundColor} />;
+            })}
+            <LabelList
+              dataKey="name"
+              position="insideLeft"
+              content={({ x, y, width, height, value }: LabelProps) => {
+                return (
+                  <text
+                    x={Number(x) + Number(width) + 2}
+                    y={Number(y) + (Number(height) || 0) / 2}
+                    fill="#000"
+                    fontSize={12}
+                    textAnchor="start"
+                  >
+                    {value}
+                  </text>
+                );
+              }}
+            />
           </Bar>
         </BarChart>
       </ResponsiveContainer>
     </TimelineWrapper>
   );
 }
-
-
-
 
 function TimelineWrapper({
   children,
@@ -88,20 +107,21 @@ function TimelineWrapper({
   children: React.ReactNode;
   shouldScroll: boolean;
 }) {
-  return shouldScroll ? (
-    <Box
-      sx={{
-        width: '100%',
-        maxWidth: '100%',
-        height: '90%',
-        overflow: 'auto',
-      }}
-    >
-      {children}
-    </Box>
-  ) : (
-    <React.Fragment>{children}</React.Fragment>
-  );
+  if (shouldScroll) {
+    return (
+      <Box
+        sx={{
+          width: '100%',
+          maxWidth: '100%',
+          height: '90%',
+          overflow: 'auto',
+        }}
+      >
+        {children}
+      </Box>
+    );
+  }
+  return <React.Fragment>{children}</React.Fragment>;
 }
 
 export default DashboardTimechart;
