@@ -1,7 +1,6 @@
 import React, { useMemo } from 'react';
-import { Link, useParams, Routes, Route, useLocation } from 'react-router-dom';
+import { Link, useParams, useLocation } from 'react-router-dom';
 import { GetDAGResponse } from '../../../models/api';
-import DAGSpecErrors from '../../../components/molecules/DAGSpecErrors';
 import DAGStatus from '../../../components/organizations/DAGStatus';
 import { DAGContext } from '../../../contexts/DAGContext';
 import DAGSpec from '../../../components/organizations/DAGSpec';
@@ -13,7 +12,7 @@ import DAGActions from '../../../components/molecules/DAGActions';
 import DAGEditButtons from '../../../components/molecules/DAGEditButtons';
 import LoadingIndicator from '../../../components/atoms/LoadingIndicator';
 import { AppBarContext } from '../../../contexts/AppBarContext';
-import useSWR, { useSWRConfig } from 'swr';
+import useSWR from 'swr';
 
 type Params = {
   name: string;
@@ -111,10 +110,6 @@ function DAGDetails() {
             <DAGEditButtons name={params.name} />
           ) : null}
         </Stack>
-
-        <Box sx={{ mt: 2, mx: 4 }}>
-          <DAGSpecErrors errors={data.Errors} />
-        </Box>
 
         <Box sx={{ mx: 4, flex: 1 }}>
           {tab == 'status' ? (
